feat: Make Validation ADT an argument of the exported function
By taking the Validation ADT as an argument, the dependency on data.validation is removed and this
package can still work with it as well as Folktale 2's Validation.
Also, removed yarn.lock file.

BREAKING CHANGE: The signature of the main (and only) function exported by this package has changed